Fixed: Property index lookup for the shell elements.
It is now placed in a new protected method since it is used
at three different places. The old implementation was buggy
as it did not properly account for indices out of range.
But the worst was that the firstEl offset was missing,
causing all stress calculations for models with beam elements
in a separate (first) patch to be wrong, unless all shells
shared common properties (not the case for the dThor models).